// angular/projekti/sebastijan-vinko-zadaca2/src/app/servisi/interfaces/DetaljiSerijeI.ts
export interface DetaljiSerijeI {
    id: number;
    originalan_naziv: string;
    naziv: string;
    opis: string;
    broj_sezona: number;
    broj_epizoda: number;
    originalan_jezik: string;
    slika: string;
    popularnost: number;
    ocjena_serije: number;
    broj_glasova: number;
    stranica: string;
}